move methods into base indy verifier
